Refactor PaletteMapper to enhance regex organization and add new methods for color path resolution. Introduced a dictionary for grouped regex patterns and improved grammar mapping logic for better maintainability and performance. Added progress bar after loading a theme for the color mapping process.
